+++ examples/demo/examples/task.rs
@@ -1,6 +1,6 @@
 use std::env;
 
-use blo::app::App;
+use demo_app::app::App;
 use loco_rs::{
     boot::{create_context, run_task},
     environment::{resolve_from_env, Environment},